change columns attr to store a hash of columns
The columns attribute will now store a hash of columns keyed by the name
of the column. The order in which the columns are in the source file is
maintained by the ordered_columns attribute, which uses the position
attribute in the columns hash to determine the sort order.

By moving to a hash, it'll allow easier access to the column attributes
for things like datatype, when I get there.