Abstract access to Text::CSV_XS
Text::xVS was *WAY* too slow when dealing with "large" CSV files
(1MB-4MB). Switching to Text::CSV_XS will provide major speed benefits,
but the interface is different, and more cumbersome to access (always
having to pass a filehandle).

This class takes away the cumbersome and the changes required for the
switch of the CSV brains.