performance improvements via limits and reorg
There are issues with the datatype detection and large files. 27,000
rows takes a long time to determine what the datatypes are. I think it
has to do with the date checking, but still need further investigation.
As a stop gap, I have included a new option to limit the sample size of
the data to 500 rows.

This doesn't impact the length determination as that's handled
separately, however it can impact the percision of numbers. Knowing the
data being parsed is key.