Use new DateTime format for parsing
Create a new parser based on a combination of the SQLite parser and
cover the gaps that other databases use.

The DateTime::Format::Natural parser does a lot of different checks
against dates and is WAY more than what's required for this project. The
SQLite parser provides the bulk of what is required with a few minor
differences.